requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cloudinary==1.42.2
Django==5.1.2
django-cors-headers==4.7.0
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.2
pillow==11.0.0
six==1.17.0
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.3.0
